Move bundledeployment underneath controllers
